# probvec: Use guvectorize with target='parallel'