This is to modify `probvec` to use [`guvectorize`](http://numba.pydata.org/numba-doc/dev/user/vectorize.html#the-guvectorize-decorator) with `target='parallel'`.
(`probvec` is mainly used to generate samples of `MarkovChain` or `DiscreteDP`.)

* Use of `guvectorize` simplifies the logic slightly; we don't need to explicitly write the outer loop.
* This version puts `sort()` in a jitted loop with nopython mode, the bottleneck of [the current `probvec`](https://github.com/QuantEcon/QuantEcon.py/blob/05d72ce1cb6a24317838909b019c9f711ccad61f/quantecon/random/utilities.py#L50).
* [Sorting by Numba may be slower than Numpy](http://numba.pydata.org/numba-doc/0.25.0/reference/numpysupported.html#other-methods), but with `target='parallel'` this version is faster than the current version on `master` for cases with large size, whereas it seems slower with small size (probably due to overhead).

Timing on a machine with 4 cores:

`master`:

```python
In [1]: import quantecon as qe

In [2]: seed = 1234

In [3]: m, k = 5000, 5000

In [4]: %timeit qe.random.probvec(m, k, random_state=seed)
1 loop, best of 3: 1.63 s per loop

In [5]: beta = 0.95

In [6]: num_states, num_actions = 500, 100

In [7]: %timeit qe.markov.random_discrete_dp(num_states, num_actions, beta, random_state=seed)
1 loop, best of 3: 1.45 s per loop
```

This version:

```python
In [1]: import quantecon as qe

In [2]: seed = 1234

In [3]: m, k = 5000, 5000

In [4]: %timeit qe.random.probvec(m, k, random_state=seed)
1 loop, best of 3: 681 ms per loop

In [5]: beta = 0.95

In [6]: num_states, num_actions = 500, 100

In [7]: %timeit qe.markov.random_discrete_dp(num_states, num_actions, beta, random_state=seed)
1 loop, best of 3: 747 ms per loop
```

This is the first case in quantecon where `guvectorize` is used; we might discuss its pros and cons here.