@spencerlyon2 2016-04-18
I'm not opposed to using guvectorize at any point in the library, but I don't love this particular change. The reason is that if I was using these tools within a larger algorithm, I would almost certainly want to parallelize a different section of my code. Having this routine always run in parallel might make that difficult.