@oyamad 2016-04-19
@spencerlyon2 We shouldn't accept this PR if it can interfere anything more important; `probvec` is just used mainly (or only) for benchmark testing.

But can you elaborate on your point? Do you have in mind any case where you may use this particular routine within a larger algorithm?