@spencerlyon2 2016-06-20
I thought about it more. I vote that we accept this. 

It would be great it numba allowed us to use `guvectorize` and then select the execution target at runtime instead of compile time, but I don't think that is possible.