@oyamad 2016-06-27
We could possibly do:

```py
def probvec(m, k, random_state=None, parallel=True):
    ...
    if parallel:
        _probvec_parallel(r, x)
    else:
        _probvec_cpu(r, x)


def _probvec(r, out):
    ...

_probvec_parallel =
    guvectorize(['(f8[:], f8[:])'], '(n), (k)', nopython=True, target='parallel')(_probvec)

_probvec_cpu =
    guvectorize(['(f8[:], f8[:])'], '(n), (k)', nopython=True, target='cpu')(_probvec)
```