@mmcky 2016-06-27
@spencerlyon2 Are you still concerned with: 

    Having this routine always run in parallel might make that difficult.

We can implement options through a wrapper (as per @oyamad solution) or make ``parallel`` the default.