@spencerlyon2 2016-06-27
I think the proposed solution sounds great. Let's go with that!
On Jun 27, 2016 5:09 AM, "mmcky" <notifications@github.com> wrote:

> @spencerlyon2 <https://github.com/spencerlyon2> Are you still concerned
> with:
>
> Having this routine always run in parallel might make that difficult.
>
> We can implement options through a wrapper (as per @oyamad
> <https://github.com/oyamad> solution) or make parallel the default.
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/QuantEcon/QuantEcon.py/pull/253#issuecomment-228774523>,
> or mute the thread
> <https://github.com/notifications/unsubscribe/AA9J3Ic1HS7bFzQg_-JSLYoWuKiFB9fXks5qP-d_gaJpZM4IJLXu>
> .
>