@mmcky 2016-07-05
@oyamad please let me know if you are happy with this update then we can merge this into the master branch. Thanks.