@oyamad 2016-07-06
@mmcky Thanks, please do merge. (I added some docstring and tests.)