@mmcky 2016-07-06
Thanks @oyamad! Merging now.